pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dvrptw-benchmark"
version = "0.1.0"
description = "Benchmark suite for static and dynamic VRPTW on Solomon RC-100 instances."
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [{name = "DVRPTW Benchmark"}]
dependencies = [
  "numpy>=1.26,<3.0",
  "scipy>=1.12,<2.0",
  "pandas>=2.2,<3.0",
  "pyarrow>=15,<20",
  "matplotlib>=3.8,<4.0",
  "plotly>=5.18,<7.0",
  "networkx>=3.2,<4.0",
  "pydantic>=2.6,<3.0",
  "hydra-core>=1.3,<2.0",
  "omegaconf>=2.3,<3.0",
  "typer>=0.12,<1.0",
  "rich>=13.7,<14.0",
  "tqdm>=4.66,<5.0",
  "ortools>=9.9,<10.0",
  "pyvrp>=0.9,<1.0",
  "torch>=2.2,<3.0",
  "lightning>=2.2,<3.0",
  "rl4co>=0.5,<1.0",
  "jupyter>=1.0,<2.0",
  "ipykernel>=6.29,<7.0",
]

[project.optional-dependencies]
gurobi = ["gurobipy>=11,<12"]
lkh3 = []
mlflow = ["mlflow>=2.12,<3.0"]
dev = ["pytest>=8.0,<9.0", "ruff>=0.3,<1.0", "mypy>=1.9,<2.0"]

[project.scripts]
dvrptw-heuristics = "dvrptw_bench.cli.heuristics_cli:app"
dvrptw-rl = "dvrptw_bench.cli.rl_cli:app"
dvrptw-hybrid = "dvrptw_bench.cli.hybrid_cli:app"
dvrptw-run-all = "dvrptw_bench.cli.run_all_cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[dependency-groups]
dev = [
    "ipykernel>=6.31.0",
]
